+++ ipa-core/src/test_fixture/logging.rs
@@ -20,8 +20,7 @@ pub fn setup() {
             use metrics_tracing_context::MetricsLayer;
             use tracing::Level;
             use tracing_subscriber::{
-                filter::Directive, fmt, fmt::format::FmtSpan, layer::SubscriberExt,
-                util::SubscriberInitExt, EnvFilter,
+                filter::Directive, fmt, layer::SubscriberExt, util::SubscriberInitExt, EnvFilter,
             };
 
             let default_directive = if let Some(crate_name) = option_env!("CARGO_CRATE_NAME") {
@@ -31,17 +30,20 @@ pub fn setup() {
                 Level::INFO.into()
             };
 
+            let fmt_layer = fmt::layer().with_test_writer();
+            #[cfg(not(feature = "step-trace"))]
+            let fmt_layer = {
+                use tracing_subscriber::fmt::format::FmtSpan;
+                fmt_layer.with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
+            };
+
             tracing_subscriber::registry()
                 .with(
                     EnvFilter::builder()
                         .with_default_directive(default_directive)
                         .from_env_lossy(),
                 )
-                .with(
-                    fmt::layer()
-                        .with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
-                        .with_test_writer(),
-                )
+                .with(fmt_layer)
                 .with(MetricsLayer::new())
                 .init();
         }
